+++ py/desidatamodel/stub.py
@@ -10,8 +10,10 @@ Generate data model files from FITS files.
 import os
 import re
 from html import escape
+from pkg_resources import resource_filename
 from astropy.io import fits
 from astropy.io.fits.card import Undefined
+from astropy.table import Table
 
 from desiutil.log import log, DEBUG
 
@@ -83,10 +85,11 @@ class Stub(DataModelUnit):
     keywords_header = ('KEY', 'Example Value', 'Type', 'Comment')
     columns_header = ('Name', 'Type', 'Units', 'Description')
 
-    def __init__(self, filename, error=False):
+    def __init__(self, filename, description_file=None, error=False):
         self.filename = None
         self.error = error
         self.headers = list()
+        self.description_file = description_file
         if isinstance(filename, (list, fits.HDUList)):
             self.nhdr = len(filename)
             for k in range(self.nhdr):
@@ -256,7 +259,16 @@ class Stub(DataModelUnit):
         ncol = hdr['TFIELDS']
         c = list()
         c.append(self.columns_header)
+        if self.description_file is not None:
+            log.info('Using column description and units from %s',
+                     self.description_file)
+            desc_data = read_column_descriptions(self.description_file)
+        else:
+            desc_data = None
+
         for j in range(ncol):
+
+            # Get units from header if possible
             jj = '{0:d}'.format(j+1)
             name = hdr['TTYPE'+jj].strip()
             ttype = fits_column_format(hdr['TFORM'+jj].strip())
@@ -269,6 +281,7 @@ class Stub(DataModelUnit):
                               bad_unit, j, hdu, self.filename)
             else:
                 units = ''
+
             # Check TCOMMnn keyword, otherwise use TTYPE comment
             # for description.
             commkey = 'TCOMM'+jj
@@ -276,6 +289,31 @@ class Stub(DataModelUnit):
                 description = escape(hdr[commkey].strip())
             else:
                 description = escape(hdr.comments['TTYPE'+jj])
+
+            # For both UNITS and DESCRIPTION, column description file trumps
+            # fits header (in case fits header was wrong and we need to
+            # override it), but at least print warning if they don't match
+
+            if desc_data is not None and name in desc_data:
+                desc_units = desc_data[name]['Units']
+                if units != '' and units != desc_units:
+                    log.warning('Overriding header units "%s" with units "%s" from %s',
+                                units, desc_units, self.description_file)
+
+                    bad_unit = self.check_unit(desc_units, error=error)
+                    if bad_unit:
+                        log.debug('Non-standard (but acceptable) units "%s" detected for column %s in %s',
+                                  bad_unit, name, self.description_file)
+
+                units = desc_units
+
+                descfile_description = escape(desc_data[name]['Description'])
+                if description != '' and description != descfile_description:
+                    log.warning('Overriding header description "%s" with description "%s" from %s',
+                                description, descfile_description, self.description_file)
+
+                description = descfile_description
+
             c.append((name, ttype, units, description))
         return c
 
@@ -654,6 +692,35 @@ def extract_keywords(hdr):
     return keywords
 
 
+def read_column_descriptions(filename):
+    """Read column descriptions csv file and return dictionary
+
+    Args:
+        filename (str): csv filename with columns NAME,TYPE,UNITS,DESCRIPTION
+
+    Returns:
+        coldesc_dict[NAME] = dict with keys TYPE, UNITS, DESCRIPTION
+    """
+
+    # Use python csv library instead of astropy Table to avoid
+    # unwanted masking of blank strings
+    import csv
+
+    with open(filename) as fp:
+        header = fp.readline().strip()
+        correct_header = 'Name,Type,Units,Description'
+        if header != correct_header:
+            raise ValueError(f'{filename} header {header} should be {correct_header}')
+
+        coldesc = dict()
+        csvreader = csv.reader(fp)
+        for row in csvreader:
+            name, dtype, units, desc = row
+            coldesc[name] = dict(Type=dtype, Units=units, Description=desc)
+
+    return coldesc
+
+
 def main():
     """Entry point for the generate_model script.
 
@@ -681,11 +748,16 @@ def main():
                         help='Set log level to DEBUG.')
     parser.add_argument('filename', help='A FITS file.', metavar='FILE',
                         nargs='+')
+    parser.add_argument("--column_descriptions",
+                        help="CSV file with column info Name,Type,Units,Description; "
+                             "default=%(default)s",
+                        default=resource_filename('desidatamodel', 'data/column_descriptions.csv'))
+
     options = parser.parse_args()
     if options.verbose:
         log.setLevel(DEBUG)
     for f in options.filename:
-        stub = Stub(f)
+        stub = Stub(f, description_file=options.column_descriptions)
         data = str(stub)
         #
         # Write the file